# backend/docs/openapi-pretty.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/groups/{groupId}" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "updateGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "group-controller" ],
        "operationId" : "deleteGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "deletedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/groups/{groupId}/settings" : {
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "updateGroupSettings",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupSettingsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/reactivate" : {
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "reactivateGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "reactivatedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/members/{userId}/promote" : {
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "promoteToAdmin",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "promotedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/members/{userId}/demote" : {
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "demoteAdmin",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "demotedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/archive" : {
      "put" : {
        "tags" : [ "group-controller" ],
        "operationId" : "archiveGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "archivedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups" : {
      "post" : {
        "tags" : [ "group-controller" ],
        "operationId" : "createGroup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GroupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/members" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroupMembers",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "group-controller" ],
        "operationId" : "addMembers",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "userIds",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }, {
          "name" : "addedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/image" : {
      "post" : {
        "tags" : [ "group-controller" ],
        "operationId" : "uploadGroupImage",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "uploadedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settlements/group/{groupId}" : {
      "get" : {
        "tags" : [ "settle-up-controller" ],
        "operationId" : "getGroupSettlements",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SettlementResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invitations/group/{groupId}" : {
      "get" : {
        "tags" : [ "invitation-controller" ],
        "operationId" : "getGroupInvitations",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InvitationResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/stats" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroupStats",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupStats"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/invitation-candidates" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getUsersForInvitation",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "searchQuery",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/balances" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroupBalances",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserBalance"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{groupId}/admins" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroupAdmins",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/user/{userId}" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getUserGroups",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GroupResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/user/{userId}/created" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "getGroupsCreatedByUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GroupResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/search" : {
      "get" : {
        "tags" : [ "group-controller" ],
        "operationId" : "searchPublicGroups",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GroupResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/expenses/group/{groupId}" : {
      "get" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "getGroupExpenses",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/groups/{groupId}" : {
      "get" : {
        "tags" : [ "super-admin-controller" ],
        "operationId" : "getGroupById",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/{groupId}/members/{userId}" : {
      "delete" : {
        "tags" : [ "group-controller" ],
        "operationId" : "removeMember",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "removedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "GroupResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "createdBy" : {
            "$ref" : "#/components/schemas/UserResponse"
          },
          "defaultCurrency" : {
            "type" : "string"
          },
          "privacyLevel" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "PRIVATE", "SECRET" ]
          },
          "invitationPolicy" : {
            "type" : "string",
            "enum" : [ "CREATOR_ONLY", "ADMIN_ONLY", "ALL_MEMBERS" ]
          },
          "allowExternalPayments" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "ARCHIVED", "DELETED" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "members" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            },
            "uniqueItems" : true
          },
          "admins" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            },
            "uniqueItems" : true
          },
          "groupImageUrl" : {
            "type" : "string"
          }
        }
      },
...
      "UserBalance" : {
        "type" : "object",
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "totalPaid" : {
            "type" : "number"
          },
          "totalOwed" : {
            "type" : "number"
          },
          "netBalance" : {
            "type" : "number"
          },
          "owedByGroup" : {
            "type" : "boolean"
          },
          "absoluteBalance" : {
            "type" : "number"
          }
        }
      },
...
    }
  }
}